requirements.txt
torch
torchvision
matplotlib
Pillow
tqdm
pandas